+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 -  Keyboard help now shows images correctly in Windows 11's "desert" high-contrast mode, by [@beyackle2](https://github.com/beyackle2) in PR [#4531](https://github.com/microsoft/BotFramework-WebChat/pull/4531)
+-  Fixes [#4501](https://github.com/microsoft/BotFramework-WebChat/issues/4501). Outgoing activities restored from chat service should be marked as sent, by [@compulim](https://github.com/compulim) in PR [#4532](https://github.com/microsoft/BotFramework-WebChat/pull/4532)
 
 ### Changed